+++ pkg/odo/cli/service/utils.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"path/filepath"
 
-	svc "github.com/openshift/odo/pkg/service"
-
 	"github.com/openshift/odo/pkg/odo/cli/component"
 	"github.com/openshift/odo/pkg/util"
 )
@@ -21,16 +19,4 @@ func validDevfileDirectory(componentContext string) error {
 			"refer %q for more information", "odo servce create -h")
 	}
 	return nil
-}
-
-// decideBackend returns the type of service provider backend to be used
-func decideBackend(arg string) ServiceProviderBackend {
-	_, _, err := svc.SplitServiceKindName(arg)
-	if err != nil {
-		// failure to split provided name into two; hence ServiceCatalogBackend
-		return NewServiceCatalogBackend()
-	} else {
-		// provided name adheres to the format <operator-type>/<crd-name>; hence OperatorBackend
-		return NewOperatorBackend()
-	}
 }